+++ web/documentserver-example/csharp-mvc/Helpers/DocumentConverter.cs
@@ -225,7 +225,7 @@ namespace OnlineEditorsExampleMVC.Helpers
         private static void ProcessResponseError(int errorCode)
         {
             var errorMessage = string.Empty;
-            const string errorMessageTemplate = "Error occurred in the ConvertService.ashx: {0}";
+            const string errorMessageTemplate = "Error occurred in the ConvertService: {0}";
 
             switch (errorCode)
             {